requirements.txt
pytest
selenium
playwright
requests
allure-pytest

